Apply review feedback: use NaturalComparer and split out Linux helpers
- Sort the drive list with SabreTools.Text.Compare.NaturalComparer instead
  of a hand-rolled comparer. This drops the previous optical-first ranking,
  so the list is now plain natural order (e.g. /dev/sg* before /dev/sr*).
- Guard the Unix append entry points with a PlatformID.Unix check.
- Move the Linux enumeration helpers and the excluded-prefix field into a
  new partial class file, Drive.Linux.cs.
- Extract UnixBlockDevice into its own file.
- Add braces to the Win32NT branch and order the excluded-prefix list
  alphabetically with a comment per entry.
