+++ MPF.Frontend.Test/DriveTests.cs
@@ -240,17 +240,17 @@ namespace MPF.Frontend.Test
 
                 var actual = Drive.EnumerateUnixFixedDevices(sysfs, "/dev");
 
-                var byPath = new Dictionary<string, Drive.UnixBlockDevice>();
+                var byPath = new Dictionary<string, UnixBlockDevice>();
                 foreach (var device in actual)
                 {
                     byPath[device.DevicePath] = device;
                 }
 
-                Drive.UnixBlockDevice sda = byPath[Path.Combine("/dev", "sda")];
+                UnixBlockDevice sda = byPath[Path.Combine("/dev", "sda")];
                 Assert.Equal(InternalDriveType.HardDisk, sda.DriveType);
                 Assert.Equal(2048L * 512L, sda.TotalSize);
 
-                Drive.UnixBlockDevice sdb = byPath[Path.Combine("/dev", "sdb")];
+                UnixBlockDevice sdb = byPath[Path.Combine("/dev", "sdb")];
                 Assert.Equal(InternalDriveType.Removable, sdb.DriveType);
                 Assert.Equal(100L * 512L, sdb.TotalSize);
             }
@@ -292,55 +292,6 @@ namespace MPF.Frontend.Test
             Directory.CreateDirectory(deviceDir);
             File.WriteAllText(Path.Combine(deviceDir, "removable"), removable);
             File.WriteAllText(Path.Combine(deviceDir, "size"), sizeSectors);
-        }
-
-        #endregion
-
-        #region CompareDrivesForDisplay
-
-        [Fact]
-        public void CompareDrivesForDisplay_SrBeforeSg_WithNumericOrder()
-        {
-            var names = new List<string?>
-            {
-                "/dev/sg10", "/dev/sr2", "/dev/sg0", "/dev/sr10",
-                "/dev/sr0", "/dev/sg1", "/dev/sr1", "/dev/sr15",
-            };
-
-            names.Sort(Drive.CompareDrivesForDisplay);
-
-            Assert.Equal(new string?[]
-            {
-                "/dev/sr0", "/dev/sr1", "/dev/sr2", "/dev/sr10", "/dev/sr15",
-                "/dev/sg0", "/dev/sg1", "/dev/sg10",
-            }, names);
-        }
-
-        [Fact]
-        public void CompareDrivesForDisplay_OpticalBeforeFixedAndRemovable()
-        {
-            var names = new List<string?>
-            {
-                "/dev/sda", "/dev/sr0", "/dev/nvme0n1", "/dev/sg0",
-            };
-
-            names.Sort(Drive.CompareDrivesForDisplay);
-
-            Assert.Equal(new string?[]
-            {
-                "/dev/sr0", "/dev/sg0", "/dev/nvme0n1", "/dev/sda",
-            }, names);
-        }
-
-        [Theory]
-        [InlineData("sr2", "sr10", -1)]
-        [InlineData("sr10", "sr2", 1)]
-        [InlineData("sr2", "sr2", 0)]
-        [InlineData("sr9", "sr10", -1)]
-        [InlineData("sda", "sdb", -1)]
-        public void NaturalCompare_TreatsDigitRunsAsNumbers(string a, string b, int expectedSign)
-        {
-            Assert.Equal(expectedSign, Math.Sign(Drive.NaturalCompare(a, b)));
         }
 
         #endregion
